fix: dataType is float not decimal

+++ resources/assets/js/page/Study/Filter/ItemFilterEdit/EditTag/statInput/index.js
@@ -21,7 +21,7 @@ const statTypes = [
     },
     {
         text: '小數',
-        value: 'decimal',
+        value: 'float',
     },
 ];
